Determine permissions for app from permissions object if 'inOrgNamespace' prop is true, otherwise determine permissions from selfsubjectaccessreview results
